assets added-separate styling and imgs merged
assets added
