+++ src/assets/img/index.js
