+++ src/assets/styles/app.js
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      alignItems: 'center',
+    },
+});
+export default styles;
