+++ src/assets/styles/boxItem.js
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    box: {
+      padding: 20,
+      backgroundColor: 'rgba(0,0,0,0.05)',
+      margin: 10,
+    },
+});
+export default styles;
