+++ src/assets/styles/detailScreen.js
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+    },
+    layout: {
+      flex: 1,
+      flexDirection: 'row',
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    box: {
+      padding: 20,
+      backgroundColor: 'rgba(0,0,0,0.05)',
+      margin: 10,
+    },
+});
+export default styles;
