+++ src/assets/styles/twoColumnView.js
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    layout: {
+      flex: 1,
+      flexDirection: 'row',
+      justifyContent: 'space-between',
+    },
+});
+export default styles;
