+++ src/components/DetailScreen.js
@@ -6,24 +6,8 @@ import {
 import { Container, Content } from 'native-base';
 
 import TwoColumnView from './common/TwoColumnView';
-import BoxItem from './common/BoxItem';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  layout: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  box: {
-    padding: 20,
-    backgroundColor: 'rgba(0,0,0,0.05)',
-    margin: 10,
-  },
-});
+import Item from './common/BoxItem';
+import styles from '../assets/styles/detailScreen';
 
 export default class DetailScreen extends Component {
 
@@ -35,7 +19,7 @@ export default class DetailScreen extends Component {
     const { items } = this.props.navigation.state.params;
     console.log(items);
     const details = items.map(item => (
-      <BoxItem
+      <Item
         key={item.id}
         onPress={() => Linking.openURL(item.url)}
         {...item}
